+++ state_linux.go
@@ -1,0 +1,6 @@
+// +build linux
+
+package specs
+
+// LinuxStateDirectory holds the container's state information
+const LinuxStateDirectory = "/run/oci/containers"
